share persistence context/EntityManager across repositories
